+++ controllers/btnServiceHybridControllers.js
@@ -1,16 +1,12 @@
-// const { query } = require("../config/dbConfig");
 const { getEpochTime, getYyyyMmDd, generateQuery } = require("../lib/utils");
-const { EKPO, BTN_SERVICE_HYBRID, BTN_LIST } = require("../lib/tableName");
-
+const { EKPO, BTN_SERVICE_HYBRID } = require("../lib/tableName");
 const { resSend } = require("../lib/resSend");
 const { APPROVED, SUBMITTED_BY_VENDOR } = require("../lib/status");
 const { create_btn_no } = require("../services/po.services");
-
-const { checkTypeArr } = require("../utils/smallFun");
-const { getQuery, query, poolClient, poolQuery } = require("../config/pgDbConfig");
+const { poolClient, poolQuery } = require("../config/pgDbConfig");
 const Message = require("../utils/messages");
-const { filesData, payloadObj } = require("../services/btnServiceHybrid.services");
-const { INSERT, ACTION_SDBG, ACTION_PBG, MID_SDBG, HR_ACTION_TYPE_WAGR_COMPLIANCE, HR_ACTION_TYPE_BONUS_COMPLIANCE, HR_ACTION_TYPE_ESI_COMPLIANCE, HR_ACTION_TYPE_PF_COMPLIANCE, HR_ACTION_TYPE_LEAVE_SALARY_COMPLIANCE } = require("../lib/constant");
+const { filesData, payloadObj, getHrDetails, getSDBGApprovedFiles, getPBGApprovedFiles, vendorDetails, getContractutalSubminissionDate, getActualSubminissionDate, checkHrCompliance, addToBTNList } = require("../services/btnServiceHybrid.services");
+const { INSERT, ACTION_SDBG, ACTION_PBG, MID_SDBG } = require("../lib/constant");
 
 const getWdcInfoServiceHybrid = async (req, res) => {
   try {
@@ -148,39 +144,6 @@ const submitBtnServiceHybrid = async (req, res) => {
   }
 };
 
-
-/**
- * INSERT DATA IN BTN LIST TABLE
- * @param {Object} client 
- * @param {Object} data 
- * @param {string} status 
- * @returns { status: boolean data: Object } || Error
- */
-const addToBTNList = async (client, data, status) => {
-  console.log("data", data);
-  try {
-    let payload = {
-      btn_num: data?.btn_num,
-      purchasing_doc_no: data?.purchasing_doc_no,
-      net_claim_amount: data?.net_claim_amount,
-      net_payable_amount: data?.net_payable_amount,
-      vendor_code: data?.vendor_code,
-      created_at: data?.created_at,
-      btn_type: data?.btn_type,
-      status: status,
-    };
-    console.log("payload", payload);
-
-    let { q, val } = generateQuery(INSERT, BTN_LIST, payload);
-    let res = await poolQuery({ client, query: q, values: val });
-    if (res.length > 0) {
-      return { status: true, data: res };
-    }
-  } catch (error) {
-    throw error;
-  }
-};
-
 /**
  * BTN CREATE INITIAL DATA SEND
  * @param {Object} req 
@@ -245,178 +208,178 @@ const initServiceHybrid = async (req, res) => {
     resSend(res, false, 500, Message.DB_CONN_ERROR, error.message, null);
   }
 }
-/**
- * GET HR UPLOADED DATA 
- * @param {Object} client 
- * @param {Object} data 
- * @returns Object || Error
- */
-const getHrDetails = async (client, data) => {
-  try {
+// /**
+//  * GET HR UPLOADED DATA 
+//  * @param {Object} client 
+//  * @param {Object} data 
+//  * @returns Object || Error
+//  */
+// const getHrDetails = async (client, data) => {
+//   try {
 
-    const actionTypeArr = [
-      HR_ACTION_TYPE_BONUS_COMPLIANCE, HR_ACTION_TYPE_WAGR_COMPLIANCE, HR_ACTION_TYPE_ESI_COMPLIANCE,
-      HR_ACTION_TYPE_PF_COMPLIANCE, HR_ACTION_TYPE_LEAVE_SALARY_COMPLIANCE
-    ];
-    const initalDataVal = data.length;
-    const placeholder = actionTypeArr.map((_, index) => `$${index + initalDataVal + 1}`).join(",");
-    const q = ` SELECT    hr.cname      AS hr_name,
-                          created_by_id AS hr_id,
-                          purchasing_doc_no,
-                          created_by_id,
-                          action_type,
-                          file_name,
-                          file_path
-                FROM      hr     AS hr_activity
-                left join pa0002 AS hr
-                ON       (
-                                    hr_activity.created_by_id = hr.pernr :: CHARACTER varying )
-                WHERE     (
-                                    hr_activity.purchasing_doc_no = $1
-                          AND       hr_activity.action_type IN (${placeholder}))`;
-    console.log("[...data, ...actionTypeArr]", q, placeholder, [...data, ...actionTypeArr]);
+//     const actionTypeArr = [
+//       HR_ACTION_TYPE_BONUS_COMPLIANCE, HR_ACTION_TYPE_WAGR_COMPLIANCE, HR_ACTION_TYPE_ESI_COMPLIANCE,
+//       HR_ACTION_TYPE_PF_COMPLIANCE, HR_ACTION_TYPE_LEAVE_SALARY_COMPLIANCE
+//     ];
+//     const initalDataVal = data.length;
+//     const placeholder = actionTypeArr.map((_, index) => `$${index + initalDataVal + 1}`).join(",");
+//     const q = ` SELECT    hr.cname      AS hr_name,
+//                           created_by_id AS hr_id,
+//                           purchasing_doc_no,
+//                           created_by_id,
+//                           action_type,
+//                           file_name,
+//                           file_path
+//                 FROM      hr     AS hr_activity
+//                 left join pa0002 AS hr
+//                 ON       (
+//                                     hr_activity.created_by_id = hr.pernr :: CHARACTER varying )
+//                 WHERE     (
+//                                     hr_activity.purchasing_doc_no = $1
+//                           AND       hr_activity.action_type IN (${placeholder}))`;
+//     console.log("[...data, ...actionTypeArr]", q, placeholder, [...data, ...actionTypeArr]);
 
-    const result = await poolQuery({ client, query: q, values: [...data, ...actionTypeArr] });
+//     const result = await poolQuery({ client, query: q, values: [...data, ...actionTypeArr] });
 
-    return result;
-  } catch (error) {
-    throw error;
-  }
-}
-/**
- * SDBG APPROVE FILE DATA
- * @param {Object} client 
- * @param {Object} data 
- * @returns  Object || Error
- */
-const getSDBGApprovedFiles = async (client, data) => {
-  try {
-    let q = `SELECT file_name as sdbg_filename FROM sdbg WHERE purchasing_doc_no = $1 and status = $2 and action_type = $3`;
-    let result = await poolQuery({ client, query: q, values: data });
-    return result;
-  } catch (error) {
-    throw error
-  }
-};
+//     return result;
+//   } catch (error) {
+//     throw error;
+//   }
+// }
+// /**
+//  * SDBG APPROVE FILE DATA
+//  * @param {Object} client 
+//  * @param {Object} data 
+//  * @returns  Object || Error
+//  */
+// const getSDBGApprovedFiles = async (client, data) => {
+//   try {
+//     let q = `SELECT file_name as sdbg_filename FROM sdbg WHERE purchasing_doc_no = $1 and status = $2 and action_type = $3`;
+//     let result = await poolQuery({ client, query: q, values: data });
+//     return result;
+//   } catch (error) {
+//     throw error
+//   }
+// };
 
-/**
- * 
- * @param {Object} client 
- * @param {Object} data 
- * @returns {Promise<Object>}
- * @throws {Error} If the query execution fails, an error is thrown.
- */
-const getPBGApprovedFiles = async (client, data) => {
-  try {
-    let q = `SELECT file_name as pbg_filename FROM sdbg WHERE purchasing_doc_no = $1 and status = $2 and action_type = $3`;
-    let result = await poolQuery({ client, query: q, values: data });
-    return result;
-  } catch (error) {
-    throw error;
-  }
-};
+// /**
+//  * 
+//  * @param {Object} client 
+//  * @param {Object} data 
+//  * @returns {Promise<Object>}
+//  * @throws {Error} If the query execution fails, an error is thrown.
+//  */
+// const getPBGApprovedFiles = async (client, data) => {
+//   try {
+//     let q = `SELECT file_name as pbg_filename FROM sdbg WHERE purchasing_doc_no = $1 and status = $2 and action_type = $3`;
+//     let result = await poolQuery({ client, query: q, values: data });
+//     return result;
+//   } catch (error) {
+//     throw error;
+//   }
+// };
 
-/**
- * GET VENDOR DETAILS
- * @param {Object} client 
- * @param {Object} data 
- * @returns {Promise<Object>}
- * @throws {Error} If the query execution fails, an error is thrown.
- */
-const vendorDetails = async (client, data) => {
-  try {
-    let q = `
-            SELECT po.lifnr       AS vendor_code,
-                   vendor_t.name1 AS vendor_name,
-                   vendor_t.email AS vendor_email,
-                   vendor_t.stcd3 AS vendor_gstno
-            FROM   ekko AS po
-                   left join lfa1 AS vendor_t
-                          ON ( po.lifnr = vendor_t.lifnr )
-            WHERE  po.ebeln = $1`;
-    let result = await poolQuery({
-      client,
-      query: q,
-      values: data,
-    });
-    return result;
-  } catch (error) {
-    throw error;
-  }
+// /**
+//  * GET VENDOR DETAILS
+//  * @param {Object} client 
+//  * @param {Object} data 
+//  * @returns {Promise<Object>}
+//  * @throws {Error} If the query execution fails, an error is thrown.
+//  */
+// const vendorDetails = async (client, data) => {
+//   try {
+//     let q = `
+//             SELECT po.lifnr       AS vendor_code,
+//                    vendor_t.name1 AS vendor_name,
+//                    vendor_t.email AS vendor_email,
+//                    vendor_t.stcd3 AS vendor_gstno
+//             FROM   ekko AS po
+//                    left join lfa1 AS vendor_t
+//                           ON ( po.lifnr = vendor_t.lifnr )
+//             WHERE  po.ebeln = $1`;
+//     let result = await poolQuery({
+//       client,
+//       query: q,
+//       values: data,
+//     });
+//     return result;
+//   } catch (error) {
+//     throw error;
+//   }
 
-}
-/**
- * ACTUAL SUBMISSION DATE
- * @param {Object} client 
- * @param {Object} data 
- * @returns {Promise<Object>}
- * @throws {Error} If the query execution fails, an error is thrown.
- */
-const getActualSubminissionDate = async (client, data) => {
-  try {
+// }
+// /**
+//  * ACTUAL SUBMISSION DATE
+//  * @param {Object} client 
+//  * @param {Object} data 
+//  * @returns {Promise<Object>}
+//  * @throws {Error} If the query execution fails, an error is thrown.
+//  */
+// const getActualSubminissionDate = async (client, data) => {
+//   try {
 
-    let a_sdbg_date_q = `SELECT actualSubmissionDate AS "PLAN_DATE", milestoneText AS "MTEXT", milestoneid AS "MID" FROM actualsubmissiondate WHERE purchasing_doc_no = $1`;
-    let a_dates = await poolQuery({
-      client,
-      query: a_sdbg_date_q,
-      values: data,
-    });
-    return a_dates;
-  } catch (error) {
-    throw error;
-  }
-}
+//     let a_sdbg_date_q = `SELECT actualSubmissionDate AS "PLAN_DATE", milestoneText AS "MTEXT", milestoneid AS "MID" FROM actualsubmissiondate WHERE purchasing_doc_no = $1`;
+//     let a_dates = await poolQuery({
+//       client,
+//       query: a_sdbg_date_q,
+//       values: data,
+//     });
+//     return a_dates;
+//   } catch (error) {
+//     throw error;
+//   }
+// }
 
-/**
- * CONTRACTUAL SUBMISSION DATE
- * @param {Object} client 
- * @param {Object} data 
- * @returns {Promise<Object>}
- * @throws {Error} If the query execution fails, an error is thrown.
- */
-const getContractutalSubminissionDate = async (client, data) => {
-  try {
+// /**
+//  * CONTRACTUAL SUBMISSION DATE
+//  * @param {Object} client 
+//  * @param {Object} data 
+//  * @returns {Promise<Object>}
+//  * @throws {Error} If the query execution fails, an error is thrown.
+//  */
+// const getContractutalSubminissionDate = async (client, data) => {
+//   try {
 
-    let c_sdbg_date_q = `SELECT PLAN_DATE as "PLAN_DATE", MTEXT as "MTEXT", MID AS "MID" FROM zpo_milestone WHERE EBELN = $1`;
-    let c_dates = await poolQuery({
-      client,
-      query: c_sdbg_date_q,
-      values: data,
-    });
-    return c_dates;
-  } catch (error) {
-    throw error;
-  }
-}
+//     let c_sdbg_date_q = `SELECT PLAN_DATE as "PLAN_DATE", MTEXT as "MTEXT", MID AS "MID" FROM zpo_milestone WHERE EBELN = $1`;
+//     let c_dates = await poolQuery({
+//       client,
+//       query: c_sdbg_date_q,
+//       values: data,
+//     });
+//     return c_dates;
+//   } catch (error) {
+//     throw error;
+//   }
+// }
 
-/**
- * CONTRACTUAL SUBMISSION DATE
- * @param {Object} client 
- * @param {Object} data 
- * @returns {Promise<Object>}
- * @throws {Error} If the query execution fails, an error is thrown.
- */
-async function checkHrCompliance(client, data) {
-  try {
-    const hrUploadedData = await getHrDetails(client, [data.purchasing_doc_no]);
-    const hrCompliancUpload = new Set([...hrUploadedData.map((el) => el.action_type)]);
-    const hrCompliances = [
-      HR_ACTION_TYPE_WAGR_COMPLIANCE,
-      HR_ACTION_TYPE_ESI_COMPLIANCE,
-      HR_ACTION_TYPE_PF_COMPLIANCE,
-    ];
+// /**
+//  * CONTRACTUAL SUBMISSION DATE
+//  * @param {Object} client 
+//  * @param {Object} data 
+//  * @returns {Promise<Object>}
+//  * @throws {Error} If the query execution fails, an error is thrown.
+//  */
+// async function checkHrCompliance(client, data) {
+//   try {
+//     const hrUploadedData = await getHrDetails(client, [data.purchasing_doc_no]);
+//     const hrCompliancUpload = new Set([...hrUploadedData.map((el) => el.action_type)]);
+//     const hrCompliances = [
+//       HR_ACTION_TYPE_WAGR_COMPLIANCE,
+//       HR_ACTION_TYPE_ESI_COMPLIANCE,
+//       HR_ACTION_TYPE_PF_COMPLIANCE,
+//     ];
 
-    for (const item of hrCompliances) {
-      if (!hrCompliancUpload.has(item)) {
-        return { success: false, msg: `Please submit ${item} to process BTN !` };
-      }
-    }
+//     for (const item of hrCompliances) {
+//       if (!hrCompliancUpload.has(item)) {
+//         return { success: false, msg: `Please submit ${item} to process BTN !` };
+//       }
+//     }
 
-    return { success: true, msg: "No milestone missing" };
-  } catch (error) {
-    return { success: false, msg: "An error occurred while checking HR compliance. Please try again later."+error.message };
-  }
-}
+//     return { success: true, msg: "No milestone missing" };
+//   } catch (error) {
+//     return { success: false, msg: "An error occurred while checking HR compliance. Please try again later."+error.message };
+//   }
+// }
 
 
 module.exports = {
